ASP.NET tool helpful for individual to create splash page to present complete online identity to any visitor.
A lot of us have multiple online profiles scattered across various
services, including Facebook, LinkedIn, Flickr, and Twitter. And one
problem we face is pulling all of this information together to build a
single on-line identity — be it for personal use, or to create a
professional on-line profile. This tool enables to quickly build a
personal and dynamic splash page that points visitors to your content
from around the web.